redirectURI와 login 페이지 사이에 새로운 페이지 추가하여 중간 과정 수정 중

+++ src/main/java/com/example/prj3be/controller/KakaoOauth.java
@@ -97,7 +97,7 @@ public class KakaoOauth implements SocialOauth {
         System.out.println("accountObject = " + accountObject);
 
         return SocialUser.builder()
-                .id((Long) jsonObject.get("id"))
+                .id(jsonObject.get("id"))
                 .name((String) profileObject.get("nickname"))
                 .email((String) accountObject.get("email"))
                 .build();
